% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid_points.R
\name{create_grid_points}
\alias{create_grid_points}
\title{Creates Grid Points that Lie within the Contiguous United States}
\usage{
create_grid_points(resolution = 50)
}
\arguments{
\item{resolution}{The resolution of the grid points}
}
\value{
A data frame containing grid points within the contiguous United
States with the first column being longitude (-180,180) and the second column
containing latitude (-90,90).
}
\description{
Given a certain resolution, a data frame containing longitudinal and
latitudinal coordinates within the contiguous (48 States) United States will
be returned.
}
\examples{
# Get a grid of points with resolution 100
pt.grid <- create_grid_points(resolution=100)
}
