% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_trends.R
\name{estimate_trends}
\alias{estimate_trends}
\title{Estimates the Average Temperature Trend Over Time for a Station}
\usage{
estimate_trends(x, station_id)
}
\arguments{
\item{x}{Data frame containing all stations and weather data}

\item{station_id}{WBANNO code for a specific station}
}
\value{
A vector containing the estimated temperature trend overtime in
units of degrees Celsius per year, the statistical significance of
that trend, and the standard error estimate of the trend.
}
\description{
Given a data frame containing all station and corresponding weather data
and a WBANNO code for a specific station, a vector of predictions for
the average temperature trend over time (degrees Celsius per year)
for the station will be returned as well as information pertaining to
the statistical significance of the trend.
}
\examples{
# Get Trend for Station NC_Asheville_13_S - WBANNO 53878
trend_dat <- estimate_trends(weather_data, 53878)
}
