% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_yearly_cycle.R
\name{estimate_yearly_cycle}
\alias{estimate_yearly_cycle}
\title{Estimates the Expected Average Temperature per Day in a Year for a Given
Station}
\usage{
estimate_yearly_cycle(x, station_id)
}
\arguments{
\item{x}{Data frame containing all stations and weather data}

\item{station_id}{WBANNO code for a specific station}
}
\value{
A data frame with the first column containing day number (1-366) and
the second column containing the expected average temperature for that day.
}
\description{
Given a data frame containing all station and corresponding weather data
and a WBANNO code for a specific station, a data frame is returned
with the first column corresponding to the day number (1-366) and the
second column containing the expected average temperature (degrees Celsius)
for that day for the station will be returned.
}
\examples{
# Estimate yearly cycle for station MS_Newton_5_ENE - WBANNO 63831
ms_yearly_cycle <- estimate_yearly_cycle(weather_data, 63831)
}
