% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_points.R
\name{interpolate_points}
\alias{interpolate_points}
\title{Interpolate Grid Points from Station Data}
\usage{
interpolate_points(x, Y, X_mod, locs_interp, X_interp)
}
\arguments{
\item{x}{Data set which contains station's spatial data
(longitude and latitude)}

\item{Y}{Response vector}

\item{X_mod}{Design matrix for model}

\item{locs_interp}{Matrix or data frame of locations in which first column is
longitudes (-180, 180) and the second column is latitudes (-90,90), output of
\code{\link{create_grid_points}}}

\item{X_interp}{Design matrix for interpolations}
}
\value{
A vector of length nrow(locs_interp) of interpolated grid points.
}
\description{
Interpolate a response value for grid points in the contiguous United
States from the spatial coordinates (longitude and latitude) of stations
as well as with custom designed design matrices.
}
\examples{
# interpolate T_DAILY_AVG for grid points within the contiguous United States
x <- weather_data[,c(3,4)]
Y <- weather_data$T_DAILY_AVG
locs_interp <- create_grid_points()
X_interp <- model.matrix(~1+Longitude+Latitude, data=locs_interp)
preds <- interpolate_points(x, Y, NULL, locs_interp, X_interp)
}
