% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gridded_interpolations.R
\name{plot_gridded_interpolations}
\alias{plot_gridded_interpolations}
\title{Creates a Plot of Interpolated Values for Spatial Data in the Mainland
United States}
\usage{
plot_gridded_interpolations(x, g, t = NULL)
}
\arguments{
\item{x}{R object containing interpolations for grid points}

\item{g}{Data frame containing grid points with first column containing
longitudes (-180,180) and second column representing latitudes (-90,90)}

\item{t}{Plot title}
}
\value{
ggplot2 plot object
}
\description{
Given an R object containing interpolated data (output of
\code{\link{interpolate_points}}), and a data frame
containing spatial data (longitudinal and latitudinal coordinates)
within the contiguous (48 States) United States a ggplot2 plot object
be returned.
}
\examples{
# Plot interpolated values of T_DAILY_AVG for mainland USA
g <- create_grid_points()
w <- weather_data[,c(3,4)]
Y <- weather_data$T_DAILY_AVG
locs_interp <- create_grid_points()
X_interp <- model.matrix(~1, data=locs_interp)
x <- interpolate_points(w, Y, NULL, locs_interp, X_interp)
g.p <- plot_gridded_interpolations(x, g)
g.p # to see plot
}
