% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_data}
\alias{station_data}
\title{Information About All Stations corresponding to US Climate Reference Network
(USCRN)}
\format{
a dataframe with 236 rows and 5 columns
\describe{
    \item{station_id}{station's WBAN number}
    \item{station_name}{station's name}
    \item{state}{station's state}
    \item{LONGITUDE}{longitude of city, in decimal degrees east}
    \item{LATITUDE}{latitude of city, in decimal degrees north}
}
}
\usage{
station_data
}
\description{
A dataset with metadata about all 236 USCRN stations
}
\keyword{datasets}
